Got sweep started

+++ sweep.h
@@ -7,9 +7,9 @@ struct QsSweep; // QsSweep is opaque
 
 /* This makes a QsSource that is a triggered sweep */
 struct QsSource *qsSweep_create(
-    float period, float level, int slope, long double holdOff,
-    long double delay,
-    struct QsSource *sourceIn, int channelNum);
-/* qsComposer_destroy() or qsSource_destroy() may be used to
- * destroy a QsSweep object. */
-
+    float period, float level, int slope, float holdOff,
+    float delay, struct QsSource *sourceIn, int channelNum);
+/* The user may use qsSource_destroy() to destroy a QsSweep
+ * object if they are so inclined.  qsSource_destroy() will
+ * call the inheriting QsSweep objects private destroy thingy.
+ */
